// legacy-structure/client/src/modules/finance/models/models.ts
export type Dictionary = {
  [key: string]: string;
};
// ...
export interface ColumnTable extends IColumn {
  field: ((row: IssuesRecord) => string) | IssuesRecordKeys;
}
// ...
export type IssuesRecordKeys =
  | 'milestones'
  | 'labels'
  | 'project'
  | 'assignedTo'
  | 'id'
  | 'number'
  | 'title'
  | 'status'
  | 'hours'
  | 'repository';
// ...
export interface IssuesRecord {
  assignedTo: {
    login: string;
  };
  description: string;
  assignedToId: number;
  number: number;
  milestones: string;
  labels: string;
  last_comment: string;
  created_at: string;
  closed_at: string;
  project: {
    title: string;
    id: number;
  };
  repository: {
    title: string;
    id: number;
  };
  id: number | '';
  title: string;
  status: string;
  hours: number | '';
  url: string;
  accountId: number;
}
// ...
export interface TransactionRecord {
  collaborator: Dictionary;
  collaboratorId: number;
  status: string;
  card_balance: number;
  cost: number;
  date: Date;
  type: string;
  idTransaction: string;
}

export type TransactionRecordKeys =
  | 'collaborator'
  | 'status'
  | 'card_balance'
  | 'cost'
  | 'date';

export interface ColumnTransactionTable extends IColumn {
  field: ((row: TransactionRecord) => string) | TransactionRecordKeys;
}